+++ CHANGELOG.md
@@ -14,6 +14,7 @@ dev
  * core: Added `ons-action-sheet-button` element.
  * core: Added new method `ons.actionSheet(...)` for creating inline `ons-action-sheet` elements.
  * core: Added new method `ons.createElement(...)` that allows creating new elements from templates or inline HTML.
+ * core: A fake device back button event is now fired on ESC press.
  * ons-splitter: Push animation.
  * ons-splitter: Reveal animation.